+++ src/Background.cpp
@@ -711,7 +711,7 @@ int BackgroundImpl::Layer::FindBGSegmentForBeat( float fBeat ) const
 
 	// assumption: m_aBGChanges are sorted by m_fStartBeat
 	int i;
-	for( i=m_aBGChanges.size()-1; i>=0; i-- )
+	for( i=static_cast<int>(m_aBGChanges.size())-1; i>=0; i-- )
 	{
 		if( fBeat >= m_aBGChanges[i].m_fStartBeat )
 			return i;
