+++ src/MemoryCardManager.cpp
@@ -678,7 +678,7 @@ void MemoryCardManager::UnmountCard( PlayerNumber pn )
 bool MemoryCardManager::PathIsMemCard( RString sDir ) const
 {
 	FOREACH_PlayerNumber( p )
-		if( !sDir.Left(MEM_CARD_MOUNT_POINT[p].size()).CompareNoCase( MEM_CARD_MOUNT_POINT[p] ) )
+		if( !sDir.Left(static_cast<int>(MEM_CARD_MOUNT_POINT[p].size())).CompareNoCase( MEM_CARD_MOUNT_POINT[p] ) )
 			return true;
 	return false;
 }
