+++ src/ScreenDimensions.cpp
@@ -39,7 +39,7 @@ float ScreenDimensions::GetScreenWidth()
 	ASSERT( fScale >= 1 );
 	// ceil causes the width to come out odd when it shouldn't.
 	// 576 * 1.7778 = 1024.0128, which is rounded to 1025. -Kyz
-	int width= std::ceil(THEME_SCREEN_WIDTH * fScale);
+	int width= static_cast<int>(std::ceil(THEME_SCREEN_WIDTH * fScale));
 	width-= width % 2;
 	return (float)width;
 }
